refactor: move scroll handling logic into util/scroll
